Switch to a uint32_t for the indices

+++ include/VGL/object.h
@@ -41,7 +41,7 @@ class Mesh {
 
     private:
         std::vector<Vertex> vertices {};
-        std::vector<uint16_t> indices {};
+        std::vector<uint32_t> indices {};
 
         VmaAllocation v_buffer_allocation{ VK_NULL_HANDLE };
